+++ changelog.md
@@ -1,5 +1,9 @@
 # Gulp-bust Changelog
 
+## 29 October 2014: _0.2.0_
+
+* Normalise paths in mappings for Windows systems.
+
 ## 22 October 2014: _0.1.1_
 
 * Added licence link to readme